+++ internal/provider/api/versions.go
@@ -6,7 +6,7 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/xanzy/go-gitlab"
+	"gitlab.com/gitlab-org/api/client-go"
 )
 
 // IsGitLabVersionLessThan is a SkipFunc that returns true if the provided version is lower then