+++ internal/provider/sdk/data_source_gitlab_projects_test.go
@@ -9,7 +9,7 @@ import (
 	"strings"
 	"testing"
 
-	gitlab "github.com/xanzy/go-gitlab"
+	gitlab "gitlab.com/gitlab-org/api/client-go"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"