+++ internal/provider/utils/license_utils_test.go
@@ -1,7 +1,7 @@
 package utils
 
 import (
-	"github.com/xanzy/go-gitlab"
+	"gitlab.com/gitlab-org/api/client-go"
 	"testing"
 )